Logging level controlled by plist
Previously, the logging level of the application was controlled via a
filed in a class.  This made changing it more difficult that it should
have been. This change moves that configuration to a Logging.plist
file (which could be configurable by environment in the future) and
causes the code to read from this location. This change introduces a
configuration() function that can be re-used by other components
wishing to read configuration in the same way.  This change also
updates the logging of various components through out the application.

[resolves #13]