Tabs for Log and Reference
This change adds two tabs to the entry point UI, one for the log, and
one  for the reference.  At the moment, there are only text labels for
these tabs.  Icons will need to be added as part of #14 and #15. As
there is currently no log functionality, the reference tab is
automatically selected on startup.  This will need to change once the
log functionality is added.

[resolves #11]