Improve fetching distilleries from CloudKit
This change improves the fetching of distilleries from CloudKit.
Previously, fetching distilleries used the convenience methods in the
CloudKit API.  These methods were limited in the number of records
that could be returned which meant that not all records were being
retrieved.  The fetch now users the operations API and can use an
optional cursor to go back and get additional records after the fetch
limit has been reached.