Restart using ReactiveCocoa
This is another restart of the project using ReactiveCocoa.  This
change was embarked on after listening to an episode of Debug with
Andy Matushak on it.  The way he talked about functional reactive
programming really put a finger on the problems that I'd been
struggling with to date.  It was enough of an epiphany that I decided
to start from scratch using these new development paradigms.  Suffice
to say, I'm quite happy with how things are going so far.