Delete Dram
This change adds the ability to delete a dram from the dram listing.
As part of this effort, the code has come more into line with the
ReactiveCocoa way of doing things.  This is still a learning
experience, and so isn't done, but we're getting there.  It also adds
the ExSwift package as a dependency.