Dram Rating
This change adds dram ratings to the New Dram interface.  You can set
the property optionally in that interface and if you do, it cannot be
edited from the drams page.  If you do not set the property, it is
editable from the drams page.  The next commit will be to save the
change if it is made in the drams page.