Save Rated Dram
Previously, if a dram was unrated, you could not add a rating for it.
This change ensures that if a dram is rated, the ability to rate it is
disabled.  If it is not rated, the ability to rate it is enabled. When
a user rates it, the value is saved and the row is reloaded,
reflecting both the new value and the inability to edit it.